requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
tda-api==1.6.0
pandas==2.1.0 